+++ lib/flop_phoenix/pagination.ex
@@ -5,63 +5,6 @@ defmodule Flop.Phoenix.Pagination do
 
   require Logger
 
-  def path_on_paginate_error_msg do
-    """
-    path or on_paginate attribute is required
-
-    At least one of the mentioned attributes is required for the pagination
-    component. Combining them will both patch the URL and execute
-    the JS command.
-
-    The :path value can be a path as a string, a
-    {module, function_name, args} tuple, a {function, args} tuple, or an 1-ary
-    function.
-
-    ## Examples
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          path={~p"/pets"}
-        />
-
-    or
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          path={{Routes, :pet_path, [@socket, :index]}}
-        />
-
-    or
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          path={{&Routes.pet_path/3, [@socket, :index]}}
-        />
-
-    or
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          path={&build_path/1}
-        />
-
-    or
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          on_paginate={JS.push("paginate")}
-        />
-
-    or
-
-        <Flop.Phoenix.pagination
-          meta={@meta}
-          path={&build_path/1}
-          on_paginate={JS.dispatch("scroll-to", to: "#my-table")}
-        />
-    """
-  end
-
   @spec default_opts() :: [Flop.Phoenix.pagination_option()]
   def default_opts do
     [